Auxiliary

GADGETPlotting.center_of_mass — Method

center_of_mass(
    position_data::Matrix{<:Real},
    mass_data::Vector{<:Real},
)::NTuple{3, Float64}

Calculate the center of mass as

$R_c = \frac{1}{M} \sum_i m_i \, r_i \, ,$

where $M = \sum_i m_i$ and $m_i$ and $r_i$ are the mass and distance from the origin of the i-th particle.

Arguments

position_data::Matrix{<:Real}: The positions of the particles.

mass_data::Vector{<:Real}: The masses of the particles.

Returns

The center of mass in the unis of position_data.

GADGETPlotting.comparison — Method

comparison(x, y; atol::Float64 = 1e-5, rtol::Float64 = 1e-5)::Bool

Determines is two elements are equal.

Arguments

x: First element to be compared.

y: Second element to be compared.

atol::Float64 = 1e-5: Absolute tolerance (for compatibility).

rtol::Float64 = 1e-5: Relative tolerance (for compatibility).

Returns

Return true if x == y.

GADGETPlotting.comparison — Method

comparison(
    x::Union{Real, AbstractArray{<:Real}, Tuple{Vararg{Real}}}, 
    y::Union{Real, AbstractArray{<:Real}, Tuple{Vararg{Real}}}; 
    atol::Float64 = 1e-5, 
    rtol::Float64 = 1e-5,
)::Bool

Determines is two numbers, numeric arrays or numeric tuples are approximately equal.

Arguments

x::Union{Real, AbstractArray{<:Real}, Tuple{Vararg{Real}}}: First element to be compared.

y::Union{Real, AbstractArray{<:Real}, Tuple{Vararg{Real}}}: Second element to be compared.

atol::Float64 = 1e-5: Absolute tolerance.

rtol::Float64 = 1e-5: Relative tolerance.

Returns

Return true if every pair of elements (X, Y) in (x, y) pass julia norm(X - Y) <= max(atol, rtol * max(norm(X), norm(Y)))`

GADGETPlotting.compute_cmdf — Method

compute_cmdf(
    mass_data::Vector{Float64},
    metallicity_data::Vector{Float64},
    max_Z::Float64,
    bins::Int64; 
    <keyword arguments>
)::NTuple{2, Vector{Float64}}

Compute the cumulative metallicity distribution function up to a metallicity max_Z.

The CMDF is calculated separating the stellar metallicity in bins windows from 0 to max_Z metallicity. Within the j-th window the CMDF is

$\sum_{i = 1}^j \frac{m_i}{M_T}$ vs. $\bar{Z}_j$

or, for x_norm = true,

$\sum_{i = 1}^j \frac{m_i}{M_T}$ vs. $\bar{Z}_j / max(\bar{Z}_j)$

where $M_T$ is the total stellar mass, $m_i$ the stellar mass of the i-th window and $\bar{Z}_j$ the mean stellar metallicity of the j-th window.

mass_data and metallicity_data must be in the same units.

Arguments

mass_data::Vector{Float64}: Masses of the particles.

metallicity_data::Vector{Float64}: Metallicities of the particles.

max_Z::Float64: Maximum metallicity up to which the profile will be calculated.

bins::Int64: Number of subdivisions of [0, max_Z] to construct the plot.

x_norm::Bool = false: If the x axis will be normalized to its maximum value.

Returns

A Tuple of two Arrays. The first with the metallicities and the second with the accumulated masses.

GADGETPlotting.deep_comparison — Method

deep_comparison(
    x::Dict, 
    y::Dict; 
    atol::Float64 = 1e-5, 
    rtol::Float64 = 1e-5,
)::Bool

Determines is two dictionaries are approximately equal.

Numeric elements are compared with comparison(), everything else with isequal().

Arguments

x::Dict: First dictionary to be compared.

y::Dict: Second dictionary to be compared.

atol::Float64 = 1e-5: Absolute tolerance for numeric elements.

rtol::Float64 = 1e-5: Relative tolerance for numeric elements.

Returns

Return true if every pair of elements pass the equality tests.

GADGETPlotting.deep_comparison — Method

deep_comparison(
    x::Union{AbstractArray, Tuple}, 
    y::Union{AbstractArray, Tuple}; 
    atol::Float64 = 1e-5, 
    rtol::Float64 = 1e-5,
)::Bool

Determines is two arrays or tuples are approximately equal.

Numeric elements are compared with comparison(), everything else with isequal().

Arguments

x::Union{AbstractArray, Tuple}: First array to be compared.

y::Union{AbstractArray, Tuple}: Second array to be compared.

atol::Float64 = 1e-5: Absolute tolerance for numeric elements.

rtol::Float64 = 1e-5: Relative tolerance for numeric elements.

Returns

Return true if every pair of elements pass the equality tests.

GADGETPlotting.density_profile — Method

density_profile(
    mass_data::Vector{Float64},
    distance_data::Vector{Float64},
    max_radius::Float64,
    bins::Int64,
)::NTuple{2, Vector{Float64}}

Compute a density profile up to a radius max_radius.

It divides the sphere of radious max_radius, centered at (0, 0, 0), in bins spherical shells of equal width max_radius / bins. This results in a density for the j-th shell of

$\rho = \frac{M}{\frac{4}{3}\,\pi\,\mathrm{width}^3\,(3\,j^2 - 3\,j + 1)}$

where $M$ is the total mass within each shell.

max_radius and distance_data must be in the same units.

Arguments

mass_data::Vector{Float64}: Masses of the particles.

distance_data::Vector{Float64}: Radial distances of the particles.

max_radius::Float64: Maximum distance up to which the profile will be calculated.

bins::Int64: Number of subdivisions of [0, max_radius] to be used for the profile.

Returns

A Tuple of two Arrays. The first with the radial distances and the second with the densities.

GADGETPlotting.energy_integrand — Method

energy_integrand(header::GadgetIO.SnapshotHeader, a::Float64)::Float64

Give the integrand of the scale factor to physical time function

$\frac{1}{H\,\sqrt{\epsilon}}$,

where $H = H_0 \, a$ and $\epsilon = \Omega_\lambda + (1 - \Omega_\lambda - \Omega_0) / a^2 + \Omega_0 / a^3$.

Arguments

header::GadgetIO.SnapshotHeader: Header of the relevant snapshot file.

a::Float64: Dimensionless scale factor.

Returns

The integrand evaluated in a in Gyr.

GADGETPlotting.format_error — Method

format_error(mean::Float64, error::Float64)::String

Format the mean and error values.

It follows the traditional rules for error presentation. The error has only one significant digit, unless such digit is a one, in which case, two significant digits are used. The mean will have a number of digits such as to match the error.

Arguments

mean::Float64: Mean value.

error::Float64: Error value. It must be positive.

Returns

A Tuple with the formatted mean and error values.

Examples

julia> format_error(69.42069, 0.038796)
(69.42, 0.04)

julia> format_error(69.42069, 0.018796)
(69.421, 0.019)

julia> format_error(69.42069, 0.0)
(69.42069, 0.0)

julia> format_error(69.42069, 73.4)
(0.0, 70.0)

GADGETPlotting.kennicutt_schmidt_law — Method

kennicutt_schmidt_law(
    gas_mass_data::Vector{Float64},
    gas_distance_data::Vector{Float64},
    temperature_data::Vector{Float64},
    star_mass_data::Vector{Float64},
    star_distance_data::Vector{Float64},
    age_data::Vector{Float64},
    temp_filter::Float64,
    age_filter::Float64,
    max_r::Float64; 
    <keyword arguments>
)::Union{Nothing, Dict{String, Any}}

Compute mass area density and the SFR area density for the Kennicutt-Schmidt law.

temp_filter and temperature_data must be in the same units, and age_filter and age_data must be in the same units too.

Arguments

gas_mass_data::Vector{Float64}: Masses of the gas particles.

gas_distance_data::Vector{Float64}: 2D distances of the gas particles.

temperature_data::Vector{Float64}: Temperatures of the gas particles.

star_mass_data::Vector{Float64}: Masses of the stars.

star_distance_data::Vector{Float64}: 2D distances of the stars.

age_data::Vector{Float64}: Ages of the stars.

temp_filter::Float64: Maximum temperature allowed for the gas particles.

age_filter::Unitful.Quantity: Maximum star age allowed for the calculation of the SFR. It should be approximately equal to the time step of the snapshots.

max_r::Float64: Maximum distance up to which the parameters will be calculated.

bins::Int64 = 50: Number of subdivisions of [0, max_r] to be used. It has to be at least 5.

Returns

A dictionary with three entries.

Key "RHO" => Logarithm of the area mass densities.

Key "SFR" => Logarithm of the SFR area densities.

Key "LM" => Linear model given by GLM.jl.

Or nothing if there are less than 5 data point in the end result.

GADGETPlotting.make_video — Method

make_video(
    source_path::String,
    source_format::String,
    output_path::String,
    output_filename::String,
    frame_rate::Int64,
)::Nothing

Make a MP4 video from a series of images.

The H.264 codec is used with no compression and the source images can be in any format available in ImageIO.jl, e.g. ".png", ".svg", ".jpeg", etc.

Arguments

source_path::String: Path to the directory containing the images.

source_format::String: File format of the source images.

output_path::String: Path to the directory where the resulting video will be saved.

output_filename::String: Name of the video to be generated without extension.

frame_rate::Int64: Frame rate of the video to be generated.

GADGETPlotting.mass_profile — Method

mass_profile(
    mass_data::Vector{Float64},
    distance_data::Vector{Float64},
    max_radius::Float64,
    bins::Int64,
)::NTuple{2, Vector{Float64}}

Compute an accumulated mass profile up to a radius max_radius.

It divides the sphere of radious max_radius, centered at (0, 0, 0), in bins spherical shells of equal width max_radius / bins. This results in a accumulated mass for the j-th shell of

$M = \sum_{i = 1}^j m_i$

where $m_i$ is the total mass within the i-th shell.

max_radius and distance_data must be in the same units.

Arguments

mass_data::Vector{Float64}: Masses of the particles.

distance_data::Vector{Float64}: Radial distances of the particles.

max_radius::Float64: Maximum distance up to which the profile will be calculated.

bins::Int64: Number of subdivisions of [0, max_radius] to be used for the profile.

Returns

A Tuple of two Arrays. The first with the radial distances and the second with the accumulated masses.

GADGETPlotting.metallicity_profile — Method

metallicity_profile(
    mass_data::Vector{Float64},
    distance_data::Vector{Float64},
    z_data::Vector{Float64},
    max_radius::Float64,
    bins::Int64,
)::NTuple{2, Vector{Float64}}

Compute a metallicity profile up to a radius max_radius, and normalize it to the solar metallicity.

It divides the sphere of radious max_radius, centered at (0, 0, 0), in bins spherical shells of equal width max_radius / bins. This results in a relative metallicity for the j-th shell of

$\rho = \frac{z}{M\,Z_\odot}$

where $M$ is the total mass and $z$ the total content of metals, within each shell.

max_radius and distance_data must be in the same units. z_data and mass_data must be in the same units.

Arguments

mass_data::Vector{Float64}: Masses of the particles.

distance_data::Vector{Float64}: Radial distances of the particles.

z_data::Vector{Float64}: Metal content of the particles in mass units.

max_radius::Float64: Maximum distance up to which the profile will be calculated.

bins::Int64: Number of subdivisions of [0, max_radius] to be used for the profile.

Returns

A Tuple of two Arrays. The first with the radial distances and the second with the metallicities.

GADGETPlotting.num_integrate — Function

num_integrate(
    func::Function, 
    inf_lim::Float64, 
    sup_lim::Float64, 
    steps::Int64 = 200,
)::Float64

Give the numerical integral of func between inf_val and sup_val.

The result is given by

$\int_\mathrm{inf\_lim}^\mathrm{sup\_lim} f(x) \mathrm{dx} \approx \sum_{i = 1}^\mathrm{steps} f(\mathrm{inf\_lim} + \mathrm{width}\,i ) \, ,$

where $\mathrm{width} = (\mathrm{sup\_lim} - \mathrm{inf\_lim}) / \mathrm{steps}$.

Arguments

func::Function: 1D function to be integrated.

inf_lim::Float64: Inferior limit of the integral.

sup_lim::Float64: Superior limit if the integral.

steps::Int64: Number of subdivisions to be used for the discretization of the sup_lim - inf_lim region.

Returns

The value of the integral.

Examples

julia> num_integrate(sin, 0, 3π)
1.9996298761360816

julia> num_integrate(x -> x^3 + 6 * x^2 + 9 * x + 2, 0, 4.69)
438.9004836958452

julia> num_integrate(x -> exp(x^x), 0, 1.0)
2.1975912134624904

GADGETPlotting.pass_all — Method

pass_all(snap_file::String, type::String)::Vector{Int64}

Default filter function for read_blocks_over_all_files.

It does not filter out any particles, allowing the data acquisition functions to gather all data.

Arguments

snap_file::String: Snapshot file path.

type::String: Particle type. "gas" -> Gas particle. "dark_matter" -> Dark matter particle. "stars" -> Star particle.

Returns

A Vector with the indices of the allowed particles.

GADGETPlotting.relative — Function

relative(
    p::Plots.Plot,
    rx::Float64,
    ry::Float64,
    rz::Union{Float64, Nothing} = nothing; 
    <keyword arguments>
)::Union{NTuple{2, Float64}, NTuple{3, Float64}}

Give the absolute coordinates within a Plot, given the relative ones.

Arguments

p::Plots.Plot: Plot for which the absolute coordinates will be calculated.

rx::Float64: relative x coordinate, rx ∈ [0, 1].

ry::Float64: relative y coordinate, ry ∈ [0, 1].

rz::Union{Float64,Nothing} = nothing: relative z coordinate, rz ∈ [0, 1].

log::Union{NTuple{2, Bool}, NTuple{3, Bool}} = (false, false, false) = If the x, y or z axis will be in a logarithmic scale.

Returns

A Tuple with the absolute coordinates: (x, y) or (x, y, z).

Examples

```julia-repl julia> GADGETPlotting.relative(plot(rand(100)), 0.5, 0.5) (50.5, 0.5047114800322484)

julia> GADGETPlotting.relative(surface(rand(100, 100)), 0.5, 0.5, 0.5) (50.5, 50.5, 0.5000284432744244)

GADGETPlotting.smooth_window — Method

smooth_window(
    x_data::Vector{<:Real},
    y_data::Vector{<:Real},
    bins::Int64,
)::NTuple{2, Vector{Float64}}

Separate the range of values of x_data in bins contiguous windows, and replaces every x and y value within the window with the mean in order to smooth out the data.

Arguments

x_data::Vector{<:Real}: Data used to create the windows.

y_data::Vector{<:Real}: Data to be smoothed out.

bins::Int64: Number of windows to be used in the smoothing.

log::Bool = false: If the x axis will be divided using logarithmic bins.

Returns

The smooth data.