Data Acquisition

GADGETPlotting.get_age — Method

get_age(snapshot::String, time::Unitful.Quantity; <keyword arguments>)::Dict{String,Any}

Get the ages of the stars at a specific time step.

Arguments

snapshot::String: Path to a given snapshot.

time::Unitful.Quantity: Clock time of snapshot, with units. All available time units in Unitful.jl and UnitfulAstro.jl can be used.

sim_cosmo::Int64 = 0: Value of the GADGET variable ComovingIntegrationOn: 0 -> Newtonian simulation (static universe). 1 -> Cosmological simulation (expanding universe).

snap_0::String = "": Path to the fist snapshot. Only relevant for cosmological simulations (sim_cosmo = 1).

filter_function::Function = pass_all: A function with the signature: foo(snap_file::String, type::String)::Vector{Int64}. See pass_all() in src/auxiliary.jl for an example. By default no particles are filtered.

Returns

A dictionary with two entries.

Key "ages" => The ages of the stars.

Key "unit" => The unit of time used.

GADGETPlotting.get_birth_place — Method

get_birth_place(
    snap_index::Int64,
    snap_files::Vector{String},
    time_stamps::Vector{Float64},
    stamps_unit::Unitful.FreeUnits; 
    <keyword arguments>
)::Dict{String, Any}

Get the birth location of the stars in a given snapshot.

Arguments

snap_index::Int64: Index in snap_files of the snapshot whose stars will be located.

snap_files::Vector{String}: Output of the function get_snapshot_paths corresponding to the key "snap_files", containing an Array with the paths to the snapshots.

time_stamps::Vector{Float64}: Clock time of every snapshot in snap_files.

stamps_unit::Unitful.FreeUnits: Unit of time of the time_stamps.

sim_cosmo::Int64 = 0: Value of the GADGET variable ComovingIntegrationOn: 0 -> Newtonian simulation (static universe). 1 -> Cosmological simulation (expanding universe).

filter_function::Function = pass_all: A function with the signature: foo(snap_file::String, type::String)::Vector{Int64}. See pass_all in src/auxiliary.jl for an example. By default no particles are filtered.

length_unit::Unitful.FreeUnits = UnitfulAstro.kpc: Unit of length to be used in the output, all available length units in Unitful.jl and UnitfulAstro.jl can be used.

time_unit::Unitful.FreeUnits = UnitfulAstro.Myr: Unit of time of time_stamps.

Returns

A 2 dimensional arrays with the positions of the stars. Each row is a star and each column corresponds to coordinates x, y and z respectively.

GADGETPlotting.get_density — Method

get_density(snapshot::String; <keyword arguments>)::Dict{String,Any}

Get the densities of the gas particles at a specific time step.

Arguments

snapshot::String: Path to the snapshot file.

sim_cosmo::Int64 = 0: Value of the GADGET variable ComovingIntegrationOn: 0 -> Newtonian simulation (static universe). 1 -> Cosmological simulation (expanding universe).

filter_function::Function = pass_all: A function with the signature: foo(snap_file::String, type::String)::Vector{Int64}. See pass_all() in src/auxiliary.jl for an example. By default no particles are filtered.

density_unit::Unitful.FreeUnits = UnitfulAstro.Msun / UnitfulAstro.kpc^3: Unit of density to be used in the output, all available density units in Unitful.jl and UnitfulAstro.jl can be used.

Returns

A dictionary with two entries.

Key "density" => Array with the densities of the gas particles.

Key "unit" => The unit of density used, i.e. is a pass-through of density_unit.

GADGETPlotting.get_hsml — Method

get_hsml(snapshot::String; <keyword arguments>)::Dict{String,Any}

Get the smoothing lengths of the gas particles at a specific time step.

Arguments

snapshot::String: Path to a given snapshot.

sim_cosmo::Int64 = 0: Value of the GADGET variable ComovingIntegrationOn: 0 -> Newtonian simulation (static universe). 1 -> Cosmological simulation (expanding universe).

filter_function::Function = pass_all: A function with the signature: foo(snap_file::String, type::String)::Vector{Int64}. See pass_all() in src/auxiliary.jl for an example. By default no particles are filtered.

length_unit::Unitful.FreeUnits = UnitfulAstro.kpc: Unit of length to be used in the output, all available length units in Unitful.jl and UnitfulAstro.jl can be used.

Returns

A dictionary with two entries.

Key "hsml" => A Vector with the smoothing lengths of the gas particles.

Key "unit" => The unit of length used, i.e. is a pass-through of length_unit.

GADGETPlotting.get_mass — Method

get_mass(snapshot::String, type::String; <keyword arguments>)::Dict{String,Any}

Get the mass of the particles at a specific time step.

Arguments

snapshot::String: Path to a given snapshot.

type::String: Particle type. "gas" -> Gas particle. "dark_matter" -> Dark matter particle. "stars" -> Star particle.

sim_cosmo::Int64 = 0: Value of the GADGET variable ComovingIntegrationOn: 0 -> Newtonian simulation (static universe). 1 -> Cosmological simulation (expanding universe).

filter_function::Function = pass_all: A function with the signature: foo(snap_file::String, type::String)::Vector{Int64}. See pass_all() in src/auxiliary.jl for an example. By default no particles are filtered.

mass_unit::Unitful.FreeUnits = UnitfulAstro.Msun: Unit of mass to be used in the output, all available mass units in Unitful.jl and UnitfulAstro.jl can be used.

Returns

A dictionary with three entries.

Key "mass" => A Vector with the masses of the particles.

Key "unit" => The unit of mass used, i.e. is a pass-through of mass_unit.

Key "type" => A String with the particle type.

GADGETPlotting.get_metallicity — Method

get_metallicity(snapshot::String, type::String; <keyword arguments>)::Dict{String,Any}

Get the metallicity (as mass content of metals) of the particles at a specific time step.

Arguments

snapshot::String: Path to a given snapshot.

type::String: Particle type. "gas" -> Gas particle. "stars" -> Star particle.

sim_cosmo::Int64 = 0: Value of the GADGET variable ComovingIntegrationOn: 0 -> Newtonian simulation (static universe). 1 -> Cosmological simulation (expanding universe).

filter_function::Function = pass_all: A function with the signature: foo(snap_file::String, type::String)::Vector{Int64}. See pass_all() in src/auxiliary.jl for an example. By default no particles are filtered.

mass_unit::Unitful.FreeUnits = UnitfulAstro.Msun: Unit of mass to be used in the output, all available mass units in Unitful.jl and UnitfulAstro.jl can be used.

Returns

A dictionary with two entries.

Key "Z" => A Vector with the metallicities of the particles.

Key "unit" => The unit of mass used, i.e. is a pass-through of mass_unit.

Key "type" => A String with the particle type.

GADGETPlotting.get_position — Method

get_position(snapshot::String; <keyword arguments>)::Dict{String, Any}

Get the coordinates of the particles at a specific time step.

Arguments

snapshot::String: Path to a given snapshot.

sim_cosmo::Int64 = 0: Value of the GADGET variable ComovingIntegrationOn: 0 -> Newtonian simulation (static universe). 1 -> Cosmological simulation (expanding universe).

filter_function::Function = pass_all: A function with the signature: foo(snap_file::String, type::String)::Vector{Int64}. See pass_all() in src/auxiliary.jl for an example. By default no particles are filtered.

box_size::Unitful.Quantity = 1000.0UnitfulAstro.kpc: Size of the plotting region if vacuum boundary conditions were used. It has to have units but they don't have to be the same as length_unit.

length_unit::Unitful.FreeUnits = UnitfulAstro.kpc: Unit of length to be used in the output, all available length units in Unitful.jl and UnitfulAstro.jl can be used.

Returns

A dictionary with six entries.

Keys "gas", "dark_matter", "stars" => 2 dimensional arrays with the positions of the particles of the type given by the key. Each row is a particle and each column correspond to coordinates x, y and z respectively.

Key "box_size" => The range of values for the plotting of the positions, i.e. a range of ± box_size if vacuum boundary conditions were used, or (0, header.boxsize) if periodic boundary conditions were used. Notice how the side length of the region is 2 * box_size for vacuum boundary conditions and header.boxsize for periodic boundary conditions.

Key "periodic" => A Boolean indicating the type of boundary condition. false -> vacuum boundary condition. true -> periodic boundary condition.

Key "unit" => The unit of length used, i.e. is a pass-through of length_unit.

GADGETPlotting.get_sfr_txt — Method

get_sfr_txt(
    source_path::String,
    snapshot::String; 
    <keyword arguments>
)::Dict{Union{Int64, String}, Any}

Get the column data from the sfr.txt file.

Transform from internal units to the ones given by mass_unit, time_unit and sfr_unit.

Warning

This function takes a modified version of sfr.txt generated by a private version of GADGET3. GADGET4 produces a sfr.txt, but it is not compatible with this function.

Arguments

source_path::String: Path to the directory containing the sfr.txt file.

snapshot::String: Path to a particular snapshot file, to use its header.

sim_cosmo::Int64 = 0: Value of the GADGET variable ComovingIntegrationOn: 0 -> Newtonian simulation (static universe). 1 -> Cosmological simulation (expanding universe).

mass_unit::Unitful.FreeUnits = UnitfulAstro.Msun: Unit of mass to be used in the output, all available mass units in Unitful.jl and UnitfulAstro.jl can be used.

time_unit::Unitful.FreeUnits = UnitfulAstro.Myr: Unit of time to be used in the output, all available time units in Unitful.jl and UnitfulAstro.jl can be used.

sfr_unit::Unitful.FreeUnits = UnitfulAstro.Msun / UnitfulAstro.yr: Unit of mass/time to be used in the output, all available time and mass units in Unitful.jl and UnitfulAstro.jl can be used.

Returns

A dictionary with six entries.

Key "1" => The first column (time).

Key "2" => The second column (total mass - probability).

Key "3" => The third column (SFR - original GADGET).

Key "4" => The fourth column (SFR - probability).

Key "5" => The fifth column (real total mass).

Key "6" => The sixth column (real SFR).

GADGETPlotting.get_snapshot_path — Method

function get_snapshot_path(
    base_name::String,
    source_path::String,
)::Dict{String, Vector{String}}

Get the paths to the GADGET output files, grouping them by snapshot.

Arguments

base_name::String: Base name of the snapshot files, set in the GADGET variable SnapshotFileBase.

source_path::String: Path to the directory containing the snapshot files, set in the GADGET variable OutputDir.

Returns

A dictionary with two entries.

Key "numbers" => A Vector with the numbers that characterize each snapshot.

Key "snap_files" => A Vector with the paths to the snapshot files.

GADGETPlotting.get_temperature — Method

get_temperature(snapshot::String; <keyword arguments>)::Dict{String,Any}

Get the temperature of the gas particles at a specific time step.

Arguments

snapshot::String: Path to a given snapshot.

sim_cosmo::Int64 = 0: Value of the GADGET variable ComovingIntegrationOn: 0 -> Newtonian simulation (static universe). 1 -> Cosmological simulation (expanding universe).

filter_function::Function = pass_all: A function with the signature: foo(snap_file::String, type::String)::Vector{Int64}. See pass_all() in src/auxiliary.jl for an example. By default no particles are filtered.

temp_unit::Unitful.FreeUnits = Unitful.K: Unit of temperature to be used in the output, all available temperature units in Unitful.jl and UnitfulAstro.jl can be used.

Returns

A dictionary with two entries.

Key "temperature" => A Vector with the temperatures of the particles.

Key "unit" => The unit of temperature used, i.e. is a pass-through of temp_unit.

GADGETPlotting.get_time_evolution — Method

get_time_evolution(snap_files::Vector{String}; <keyword arguments>)::Dict{String, Any}

Get several parameters defined at every snapshot, as a series of values for the whole simulation.

The parameters are:

"scale_factor"

"redshift"

"clock_time" (Physical time)

"sfr" (SFR)

"sfr_prob" (SFR probability - Not normalized)

"gas_number" (Gas particle number)

"dm_number" (Dark matter particle number)

"star_number" (Star number)

"gas_mass" (Total gas mass)

"dm_mass" (Total dark matter mass)

"star_mass" (Total star mass)

"gas_density" (Global gas density)

"gas_frac" (Gas fraction)

"dm_frac" (Dark matter fraction)

"star_frac" (Star fraction)

"gasbarfrac" (Baryonic gas fraction)

"starbarfrac" (Baryonic star fraction)

Arguments

snap_files::Vector{String}: Output of the function get_snapshot_paths corresponding to the key "snap_files", containing an Array with the paths to the snapshots.

sim_cosmo::Int64 = 0: Value of the GADGET variable ComovingIntegrationOn: 0 -> Newtonian simulation (static universe). 1 -> Cosmological simulation (expanding universe).

filter_function::Function = pass_all: A function with the signature: foo(snap_file::String, type::String)::Vector{Int64}. See pass_all() in src/auxiliary.jl for an example. By default no particles are filtered.

mass_unit::Unitful.FreeUnits = UnitfulAstro.Msun: Unit of mass to be used in the output, all available mass units in Unitful.jl and UnitfulAstro.jl can be used.

time_unit::Unitful.FreeUnits = UnitfulAstro.Myr: Unit of time to be used in the output, all available time units in Unitful.jl and UnitfulAstro.jl can be used.

sfr_unit::Unitful.FreeUnits = UnitfulAstro.Msun / UnitfulAstro.yr: Unit of mass/time to be used in the output, all available time and mass units in Unitful.jl and UnitfulAstro.jl can be used.

length_unit::Unitful.FreeUnits = UnitfulAstro.kpc: Unit of length to be used in the output, all available length units in Unitful.jl and UnitfulAstro.jl can be used.

density_unit::Unitful.FreeUnits = UnitfulAstro.Msun / UnitfulAstro.kpc^3: Unit of density to be used in the output, all available density units in Unitful.jl and UnitfulAstro.jl can be used.

Returns

A dictionary.

Key "{property}" => A Vector with the numeric values of the {property} in the key (one value per snapshot) for the whole simulation.

Key "units" => A dictionary with the units used, for easy piping with other functions.

Key "labels" => A dictionary with the labels to be used when plotting the quantities.